refactor: condition 다시 query로 수정
