+++ src/main/java/com/trustamarket/paymentservice/paymentservice/application/dto/query/PaymentDetailQuery.java
@@ -1,0 +1,8 @@
+package com.trustamarket.paymentservice.paymentservice.application.dto.query;
+
+import java.util.UUID;
+
+public record PaymentDetailQuery(
+        UUID paymentId,
+        UUID userId
+) {}
