+++ src/main/java/com/trustamarket/paymentservice/paymentservice/domain/repository/PaymentRepository.java
@@ -1,6 +1,6 @@
 package com.trustamarket.paymentservice.paymentservice.domain.repository;
 
-import com.trustamarket.paymentservice.paymentservice.domain.repository.condition.PaymentSearchCondition;
+import com.trustamarket.paymentservice.paymentservice.application.dto.query.PaymentSearchQuery;
 import com.trustamarket.paymentservice.paymentservice.domain.entity.Payment;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
@@ -10,5 +10,5 @@ import java.util.UUID;
 public interface PaymentRepository {
     Payment saveAndFlush(Payment payment);
     Payment findById(UUID paymentId);
-    Slice<Payment> searchPayments(PaymentSearchCondition query, Pageable pageable);
+    Slice<Payment> searchPayments(PaymentSearchQuery query, Pageable pageable);
 }
