+++ src/main/java/com/trustamarket/paymentservice/paymentservice/presentation/PaymentController.java
@@ -5,7 +5,7 @@ import com.trustamarket.common.response.SlicedResponse;
 import com.trustamarket.common.util.SecurityUtil;
 import com.trustamarket.paymentservice.paymentservice.application.dto.command.FailPaymentCommand;
 import com.trustamarket.paymentservice.paymentservice.application.dto.query.PaymentDetailQuery;
-import com.trustamarket.paymentservice.paymentservice.domain.repository.condition.PaymentSearchCondition;
+import com.trustamarket.paymentservice.paymentservice.application.dto.query.PaymentSearchQuery;
 import com.trustamarket.paymentservice.paymentservice.application.dto.command.SucceededPaymentCommand;
 import com.trustamarket.paymentservice.paymentservice.application.dto.result.FailPaymentResult;
 import com.trustamarket.paymentservice.paymentservice.application.dto.result.PaymentDetailResult;
@@ -92,7 +92,7 @@ public class PaymentController {
     ){
         UUID userId =  SecurityUtil.getCurrentUserIdOrThrow();
 
-        PaymentSearchCondition query = new PaymentSearchCondition(
+        PaymentSearchQuery query = new PaymentSearchQuery(
                 request.paymentId(),
                 userId,
                 request.minAmount(),
